+++ sporttracker/blueprints/Achievements.py
@@ -1,8 +1,8 @@
 import logging
 from datetime import datetime
 
-from flask import Blueprint, render_template
-from flask_babel import gettext
+from flask import Blueprint, render_template, url_for
+from flask_babel import gettext, format_datetime
 from flask_login import login_required
 
 from sporttracker.helpers import Helpers
@@ -43,6 +43,15 @@ def construct_blueprint():
                     ).format(currentStreak=streak[1], maxStreak=streak[0]),
                 )
             )
+
+            longestDistance = 0.0
+            longestTrackDate = gettext('no date')
+            longestTrack = AchievementCalculator.get_track_with_longest_distance_by_type(trackType)
+            if longestTrack is not None:
+                longestDistance = longestTrack.distance / 1000
+                longestTrackDate = format_datetime(longestTrack.startTime, format='dd.MM.yyyy')
+                longestTrackDate = f'<a href="{url_for("maps.showSingleTrack", track_id=longestTrack.id)}">{longestTrackDate}</a>'
+
             achievementList.append(
                 Achievement(
                     icon='route',
@@ -50,12 +59,10 @@ def construct_blueprint():
                     color=trackType.border_color,
                     title=gettext('Longest Track'),
                     description=gettext(
-                        'You completed <span class="fw-bold">{longestTrack} km</span> in one trip!'
+                        'You completed <span class="fw-bold">{longestTrackDistance} km</span> in one trip on <span class="fw-bold">{longestTrackDate}</span>!'
                     ).format(
-                        longestTrack=Helpers.format_decimal(
-                            AchievementCalculator.get_longest_distance_by_type(trackType),
-                            decimals=2,
-                        )
+                        longestTrackDistance=Helpers.format_decimal(longestDistance, decimals=2),
+                        longestTrackDate=longestTrackDate,
                     ),
                 )
             )